Remove Eureka-based service discovery and related configurations.
This commit eliminates all dependencies, configurations, and resources related to Eureka from the project. It includes the removal of the Discovery service, Eureka client libraries from microservices, and related settings in the `application.yml` files. These changes simplify the architecture, transitioning to a non-Eureka-based setup.

NOTE: Chaos Tests Are Broken at this point.
(This is to prepare for k8s)